local laplacian: refactor into .c/.h
